src/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.3.2
pymongo==4.6.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.0
langgraph==0.0.20
yfinance==0.2.28
python-multipart==0.0.6
python-dotenv==1.0.0
